move fetch out of the componenet
