+++ .env.local.template
@@ -1,0 +1,1 @@
+OPEN_API_KEY=[API_KEY]
