+++ app/page.tsx
@@ -2,15 +2,67 @@
 
 import { useState, createContext } from 'react';
 import { Chat } from './components/chat';
-import { Graph } from './components/model';
+import { Graph, Node } from './components/model';
 import { Github, HomeIcon } from 'lucide-react';
 import Link from 'next/link';
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
 import { CodeGraph } from './components/code-graph';
+import { toast } from '@/components/ui/use-toast';
 
 export default function Home() {
 
   const [graph, setGraph] = useState(Graph.empty());
+
+  function onFetchGraph(url: string) {
+    let value = url;
+    if (!value || value.length === 0) {
+      value = 'https://github.com/falkorDB/falkordb-py';
+    }
+
+    fetch('/api/repo', {
+      method: 'POST',
+      body: JSON.stringify({
+        url: value
+      })
+    }).then(async (result) => {
+      if (result.status >= 300) {
+        throw Error(await result.text())
+      }
+      return result.json()
+    }).then(data => {
+      let graph = Graph.create(data);
+      setGraph(graph);
+    }).catch((error) => {
+      toast({
+        variant: "destructive",
+        title: "Uh oh! Something went wrong.",
+        description: error.message,
+      });
+    });
+  }
+
+
+  function onFetchNode(node: Node) {
+    // var node: Node = evt.target.json().data;
+
+    fetch(`/api/repo/${graph.Id}/${node.name}`, {
+        method: 'GET'
+    }).then(async (result) => {
+        if (result.status >= 300) {
+            throw Error(await result.text())
+        }
+        return result.json()
+    }).then(data => {
+        graph.extend(data)
+        setGraph(graph)
+    }).catch((error) => {
+        toast({
+            variant: "destructive",
+            title: "Uh oh! Something went wrong.",
+            description: error.message,
+        })
+    })
+}
 
   return (
     <main className="h-screen flex flex-col">
@@ -30,7 +82,7 @@ export default function Home() {
 
       <PanelGroup direction="horizontal" className="w-full h-full">
         <Panel defaultSize={75} className="flex flex-col border" collapsible={true} minSize={30}>
-          <CodeGraph graph={graph}  setGraph={setGraph}/>
+          <CodeGraph graph={graph} onFetchGraph={onFetchGraph} onFetchNode={onFetchNode} />
         </Panel>
         <PanelResizeHandle className="w-1 bg-gray-500" />
         <Panel className="flex flex-col border" defaultSize={25} collapsible={true} minSize={10}>
